Setup screen brightness script

+++ install.sh
@@ -22,8 +22,38 @@ then
 	rm -r fasd/
 fi
 
-echo "Setting up the screen brightness script"
-ln -s $scriptdir/brightness.sh ~/.i3/brightness
+
+# rclocal=$(cat <<'END_HEREDOC'
+# #!/bin/sh -e
+# #
+# # rc.local
+# #
+# # This script is executed at the end of each multiuser runlevel.
+# # Make sure that the script will "exit 0" on success or any other
+# # value on error.
+# #
+# # In order to enable or disable this script just change the execution
+# # bits.
+# #
+# # By default this script does nothing.
+
+# chown pedro:users /sys/class/backlight/radeon_bl0/brightness
+
+# exit 0
+# END_HEREDOC
+# )
+
+# echo "$rclocal"
+
+if [ "$(exists i3)" -eq 1 ]
+then
+	echo "Setting up the screen brightness script"
+	ln -s $scriptdir/brightness.sh ~/.i3/brightness
+
+# 	sudo mv /etc/rc.local /etc/rc.local_old
+# 	sudo sed -i -e '$i chown pedro:users /sys/class/backlight/radeon_bl0/brightness\n' "/etc/rc.local"
+# 	# bash -c "echo "$rclocal" > "/etc/rc.local""
+fi
 
 echo "Deleting the old files"
 mv ~/.zshrc ~/.zshrc_old
